+++ src/main/java/org/rsna/video/DataChunkOutputStream.java
@@ -149,8 +149,8 @@ public class DataChunkOutputStream extends FilterOutputStream {
      * @throws java.io.IOException
      */
     public void writeShort(int v) throws IOException {
-        out.write((int) ((v >>> 0) & 0xff));
-        out.write((int) ((v >> 8) & 0xff));
+        out.write((v >>> 0) & 0xff);
+        out.write((v >> 8) & 0xff);
         incCount(2);
     }
 
@@ -167,8 +167,8 @@ public class DataChunkOutputStream extends FilterOutputStream {
     }
 
     public void writeUShort(int v) throws IOException {
-        out.write((int) ((v >>> 0) & 0xff));
-        out.write((int) ((v >> 8) & 0xff));
+        out.write((v >>> 0) & 0xff);
+        out.write((v >> 8) & 0xff);
         incCount(2);
     }
